Change IndexOutOfBounds to use i64 for negative index support

+++ crates/cctr-expr/src/lib.rs
@@ -207,8 +207,8 @@ pub enum EvalError {
         expected: usize,
         got: usize,
     },
-    #[error("index out of bounds: {index} >= {len}")]
-    IndexOutOfBounds { index: usize, len: usize },
+    #[error("index out of bounds: {index} (len: {len})")]
+    IndexOutOfBounds { index: i64, len: usize },
     #[error("key not found: {0}")]
     KeyNotFound(String),
 }
